+++ pdns/dnsdist-rings.cc
@@ -111,9 +111,9 @@ std::unordered_map<int, vector<boost::variant<string,double>>> Rings::getTopBand
 	       });
   std::unordered_map<int, vector<boost::variant<string,double>>> ret;
   uint64_t rest = 0;
-  unsigned int count = 1;
+  int count = 1;
   for(const auto& rc : rcounts) {
-    if(count==numentries+1) {
+    if (count == static_cast<int>(numentries + 1)) {
       rest+=rc.first;
     }
     else {
